DROID-3013 move objects ui to ui-core
