+++ feature-date/src/main/java/com/anytypeio/anytype/feature_date/ui/models/DateEvent.kt
@@ -3,8 +3,8 @@ package com.anytypeio.anytype.feature_date.ui.models
 import com.anytypeio.anytype.core_models.TimeInMillis
 import com.anytypeio.anytype.core_models.multiplayer.SpaceSyncAndP2PStatusState
 import com.anytypeio.anytype.core_models.primitives.TimestampInSeconds
+import com.anytypeio.anytype.core_ui.lists.objects.UiObjectsListItem
 import com.anytypeio.anytype.feature_date.viewmodel.UiFieldsItem
-import com.anytypeio.anytype.feature_date.viewmodel.UiObjectsListItem
 
 sealed class DateEvent {
